+++ src/location/location.controller.ts
@@ -3,14 +3,18 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
   Param,
   Post,
   UseGuards,
 } from '@nestjs/common';
 import { Role } from '@prisma/client';
 import { Roles } from 'src/common/decorator/role.decorator';
+import { UserDec } from 'src/common/decorator/user.decorator';
 import { LocationDto } from 'src/common/dtos/location.dto';
 import { RolesGuard } from 'src/common/gurds/role.gurds';
+import { AccessValidator } from 'src/common/utils/access-validator';
+import { UserEntitiesByType } from 'src/common/utils/user-entitites';
 import { LocationService } from './location.service';
 
 @Controller('v1/location')
@@ -30,14 +34,27 @@ export class LocationController {
   }
 
   @Get('/')
-  async getAll() {
-    return this.locationService.getAll();
+  @Roles(Role.SUPER_ADMIN, Role.ADMIN)
+  @UseGuards(RolesGuard)
+  async getAll(@UserDec() user) {
+    if (user.role == Role.SUPER_ADMIN) {
+      return this.locationService.getAllLocations();
+    }
+    const userEntitiesIds = UserEntitiesByType(
+      user?.AdminAccessMap,
+      'LOCATION',
+    );
+
+    return this.locationService.getAllUserLocation(userEntitiesIds);
   }
 
   @Delete('/:id')
-  @Roles(Role.SUPER_ADMIN)
+  @Roles(Role.SUPER_ADMIN, Role.ADMIN)
   @UseGuards(RolesGuard)
-  async deleteOne(@Param('id') id: number) {
-    return this.locationService.deleteOne(+id);
+  async deleteOne(@UserDec() user, @Param('id') id: number) {
+    if (AccessValidator(user.role, user?.AdminAccessMap, +id, 'LOCATION')) {
+      return this.locationService.deleteOne(+id);
+    }
+    throw new HttpException('forbidden', 403);
   }
 }
